Extract the kubernetes connection related data to a struct
To preserve the existing code base, we embed the new type within
the KubernetesIPAM (old type).

In a follow-up commit, this new struct will be used on a separate
golang binary (which doesn't require the KubernetesIPAM
configuration), to provide a simple connection to the Kubernetes
backend (does not support etcd).

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>