e2e tests: make WB play along with containerd
containerd sends CNI CMD check instructions to the plugins, and
whereabouts does not implement that command, returning with an
error.

As per the CNI specification - [0] - the runtime "may assume that
a failed CHECK means the container is permanently in a misconfigured
state." and as a results replies with a CNI del.

By instructing the CRI to not send CNI checks we work around this
issue.

[0] - https://github.com/containernetworking/cni/blob/master/SPEC.md#check-check-containers-networking-is-as-expected

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>