Introduce 'RangeConfiguration' type and 'IPRanges'
Introduce new type 'RangeConfiguration' for encapsulating the fields
related to the range from which whereabouts will choose a free IP to be
assigned to any pod.

Add a new field 'IPRanges' to the struct 'IPAMConfig' which is a array
of 'RangeConfiguration'. Whereabouts will assign one free IP from evey
range in 'IPRanges' to the pod.

Signed-off-by: Ayush Patel <patel.ayush08@gmail.com>